refactor(product): finish chat loading hero exit-hold extraction
Complete the extraction: fix the moved test's import path and stage
ChatView.tsx's call-site-only shrink plus the new hook file (both were
left out of the previous commit by a failed multi-path git add).
